HH member form refactor
