+++ src/Components/Steps/HouseholdMembers/utils/constants.ts
@@ -10,3 +10,14 @@ export const BASIC_INFO_GRID_STYLES: SxProps<Theme> = {
 
 // Frequency order from least to most frequent (business logic, not styling)
 export const FREQUENCY_ORDER = ['once', 'yearly', 'monthly', 'semimonthly', 'biweekly', 'weekly', 'hourly'];
+
+// Error section mapping for scroll-to-error functionality
+export const ERROR_SECTION_MAP = [
+  { key: 'birthMonth', id: 'basic-info-section' },
+  { key: 'birthYear', id: 'basic-info-section' },
+  { key: 'relationshipToHH', id: 'basic-info-section' },
+  { key: 'healthInsurance', id: 'health-insurance-section' },
+  { key: 'conditions', id: 'conditions-section' },
+  { key: 'hasIncome', id: 'income-section' },
+  { key: 'incomeStreams', id: 'income-section' },
+] as const;
